pyproject.toml
[tool.poetry]
name = "ballsdex"
version = "1.0.0a1"
description = ""
authors = ["laggron42 <laggron42@gmail.com>", "millipede <chris@millipede.eu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11.0"

# asyncio
uvloop = { version = "^0.17.0", markers = "sys_platform != 'win32'" }

# discord
"discord.py" = "^2.2.2"

# fastapi
fastapi = "^0.95.0"
fastapi-admin = {git = "https://github.com/fastapi-admin/fastapi-admin", rev = "5f34bf0d7356afa4c6034a7ab3942ea843d3b8ce"}
uvicorn = "^0.21.1"

# database ORM
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
redis = "^4.5.3"

# misc
rich = "^12.6.0"
python-dateutil = "^2.8.2"
Pillow = "^9.4.0"
aerich = "^0.6.3"
pyyaml = "^6.0"
cachetools = "^5.3.1"
pilmoji = "^2.0.3"
numpy = "^1.25.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.1"
black = {version = "^23.1.0", allow-prereleases = true}


[tool.poetry.group.metrics.dependencies]
prometheus-client = "^0.16.0"

[tool.aerich]
tortoise_orm = "ballsdex.__main__.TORTOISE_ORM"
location = "./migrations"
src_folder = "./ballsdex"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
    line-length = 99
    target-version = ['py38']
    include = '\.py$'

[tool.flake8]
    ignore = "W503,E203"
    max-line-length = 99

#[[tool.poetry.source]]
#name = 'aliyun.mirrors'
#url = 'https://mirrors.aliyun.com/pypi/simple/'
#priority = "default" 
